sv/requirements.txt
colorama==0.4.6
vunit_hdl==4.7.0
